Upgrade releases for CF release 245
This includes upgrading to etcd 75. CF 245 soft recommends v73, but this
has a known issue. We therefore went to the latest release, which
includes a fix for this, and no other functional changes.

This includes removing our fork of the routing release as the fix is now
included upstream. This reverts f9b9b0c.

This also includes removing the deployment of a separate version of uaa.
This was originially done in 9be1594 to fix a CVE. Now that we're
upgrading CF, we can go back to using the version that it ships.