Update manifests with new upstream etcd config.
There have been a few changes to how etcd is configured in the upstream
manifests. The main details are:

* etcd is now registered in consul, and accessed via that DNS name
* TLS support is enabled for etcd.

We were already registering etcd in consul (under the `etcd` name as
opposed to `cf-etcd` that upstream are using). This therefore updates
the manifest to define the properties in a manner that matches upstream.

Migrating from a non-TLS etcd config to a TLS config is a non-trivial
process[1], so we've decided to defer that migration until later. This
therefore adds all the new properties to the manifest, but leaves TLS
disabled and leaves the cert properties blank.

See
https://github.com/cloudfoundry/cf-release/commit/c432b73#diff-b9982126ee4438e20576a100670694d8
and https://github.com/cloudfoundry/cf-release/commit/22fc918 for
details of the upstream changes.

[1]https://docs.google.com/document/d/1ZzWzp3H6H3t1ikk6Fl-8x1LX2a_0dHPJ5MMLEwY0inI/edit