Revert "Merge pull request #495 from alphagov/feature/124720541-add-metron_agent-to-all-vms"
This reverts commit 7cac814425c860985acb2b9a35259954c175e4e4, reversing
changes made to b5569530477322b39f36c6af23640a2ee458f9f0.

We have added metron_agent as a shortcut to getting rsyslogd configured
for log forwarding.  This in hindsight was a mistake, as it cannot be
configured to just forward, so we have extra broken instances of the
metron_agent service.

Here we revert the addition so we can reimplement this using a different
release.