Rename ENABLE_HEALTHCHECK_DB switch
The logic was that when this was set to false, then don't deploy. But when
it was set to empty (no value), then that still meant yes.

Turn this into DISABLE_HEALTHCHECK_DB, where if set to true it means you
want to disable it. Otherwise (empty or false) it means you don't want to
disable...