Rename ENABLE_ACCEPTANCE_TESTS switch
The logic was that when this was set to false, then don't run tests. But
when it was set to empty (no value), then that meant yes.

Turn this into DISABLE_ACCEPTANCE_TESTS, where if set to true it means you
want to disable it. Otherwise (empty or false) it means you don't want
to disable...