Upload all stemcells defined in the CF manifest
We want to be able to define several stemcells in our manifest, and
be able to use then.

For that we change the task get-and-upload-stemcell so that will upload all
the stemcells defined in the manifest. We also change it to read the
stemcells from the rendered manifest, instead of the definition.