Add job definition cell_3_3
Add job definition `cell_3_3` with `instances: 0` which uses the stemcell
with  kernel 3.3 and the garden-linux release.

We need this definition to prevent the bosh-clean job from deleting
releases and stemcells that are not installed.

It would also allow us to do a alternative way to rollback to kernel
3.3 and garden-linux, by increasing the number of instances of `cell_3_3`
and later decreasing the ones in `cell`

In order to define the new `cell_3_3` job with the same config than
`cell`, we use the merge capabilities of spruce, injecting all the
configuration from cell and overriding the values that change.