Upgrade ETCD to v78
v78 fixes issue with ETCD network debugging being turned on all the time and
being undisableable. This is a new property/functionality that was not present
in previous ETCD release we were using in cf v240. Because of a bug in v75,
it produces extreme amount of logs, quadrupling total amount of all logs from
whole platform. Latest ETCD release contains fix for that and only runs this
functionality for 5 minutes from just before ETCD restart.