Add 100G of ephemeral disk for the Cells
The cells are using the local VM space for ephemeral storage, and there
is a risk that they running out of space. We should add more ephemeral
space, but there was a concern about the user impact.

After upgrade the CPI, we know that all the ephemeral disks will be EBS[1]

* Use EBS disks for ephemeral disk (/var/vcap/data) instead of instance storage
  * Very convenient when shutting down and starting up large number of machines directly in AWS
  * Can opt-in to use instance storage via  ephemeral_disk.use_instance_storage=true

In more detail, as described in [2]

Currently our cell configuration is to use ephemeral disks. The instance type
we use (r3.xlarge) has 80GB ephemeral disk. 30GB is used for swap, which
leaves 50GB available for tenant instances and other cf data.

As it turns out, this might be not enough because:

 * default droplet disk limit is 2GB. It is rarely suggested to use less
   (in docs etc.), set by the cc.maximum_app_disk_in_mb
 * available disk space is not checked as rigorously as available memory when
   CF/Diego when picking cell to host new instance It is possible that a
   eployment will be attempted on a cell that has not enough disk space available
 * CF has a cache directory in `/var/vcap/data/executor_cache`. We've seen
   this take 9G (or 20% of available space) on a cell in staging.
   We also have seen very high disk usage on cells in CI.

In this commit we add dedicated ephemeral disk. We consider that 100GB should
be able to support our current load without issues (100-150 instances)

[1] https://github.com/cloudfoundry-incubator/bosh-aws-cpi-release/releases/tag/v53
[2] https://www.pivotaltracker.com/n/projects/1275640/stories/133104393