Update plan of existing healthcheck instances
The change to the plan used for the healthcheck-db won't take effect
because the DB already exists in the persistent environments. This adds
a temporary script to check for the plan it's using, and re-create it if
it's using the old plan.

This commit can be reverted once the DB has been updated in all
persistent environments.