Use the 'Free' RDS plan for the tests where possible.
This will be cheaper, and faster (due to snapshots being disabled).

The tests that cover the final snapshots need to use a plan that has
these snapshots enabled. This therefore refactors the tests a little so
that these tests are confined to a single block, and the other tests can
therefore use the Free instances. The snapshotting tests are now also
using the small plan instead of the medium one to save cost. They were
only using the medium plan because that was the only one that existed
when these tests were created.