Explicitly set EBS ephemeral disk to all VM types
As it turns out, if you don't specify the ephemeral disk, the new
bosh will automatically create the EBS volume volume matching the
previous free size in the instance disk.

For instance, in m3.large it will create a 4 GB EBS volume by default
which matches the previous disk.

We think it is convenient to explicitly define the ephemeral disk for
every VM type.

We add a test to check that we define, at least, 10GB disk for all the
VM types. We add the ephemeral disk for ingestor, access, and
compilation types. Compilation has currently 12GB, we decided to set
20G.