Make references use the new bosh-job tag instead
With the new features in datadog agent, we don't need to create a
`job` tag, as `bosh-job` is provided to us by bosh.

We refactor all the graphs and monitors, to use the new tag instead.

Note: This does not apply for the metrics reported by datadog-nozzle,
which will be tagged with `deployment` and `job` instead.