Upload test artifacts to S3
We want to store the CF_TRACE output from tests run in the pipeline,
as they help with debugging test failures. We do not want them output
in the pipeline, as we cannot guarantee they won't contain secrets,
so we upload them to an S3 bucket.

The paths used to store the files are based from the way the upstream
cf-release does it. For the acceptance tests the path is hard coded[1].
For the smoke-tests they do not support configuring this value[2][3],
so a path consistent with that used by the acceptance tests was chosen.
These paths are absolute, and Concourse does not allow you to declare
an output with an absolute path. For example, when declaring this output:

```
outputs:
- name: artifacts
  path: /var/vcap/sys/log/acceptance_tests
```

The directory will be created under:

```
/tmp/build/$UUID/var/vcap/sys/log/acceptance_tests
```

The solution is to create a symbolic link to our artifacts directory,
so when the tests try to store files under /var/vcap/sys/log they are
exportable to the next Concourse task. This is consistent to how we
have handled similar issues when running the smoke tests[4].

Shell check error SC2035 was encountered when using TAR to create the
archive, so we use `--` to stop parsing flags before the glob[5].

[1] https://github.com/cloudfoundry/cf-release/blob/v245/jobs/acceptance-tests/templates/config.json.erb#L24
[2] https://github.com/cloudfoundry/cf-release/blob/v245/jobs/smoke-tests/templates/config.json.erb#L20-L37
[3] https://github.com/cloudfoundry/cf-release/issues/1089
[4] https://github.com/alphagov/paas-cf/blob/staging-0.0.314/concourse/tasks/smoke-tests-run.yml#L21-L24
[5] https://github.com/koalaman/shellcheck/wiki/Sc2035