Introduce dashboard for PaaS monitor overview
We are using [Smashing](https://github.com/Dashing-io/smashing) to provide a dashboard.
We chose this as it is already in use by Verify, it seems simplest and easiest out of
existing tools, and it is easy to read at a distance.

We only use the 'Health' widget at the moment, but there are plenty of other useful
integrations (eg for pivotal, pingdom, github and so on we might add later).

We include a readme describing the expected usage, which we will implement later.