Use latest aws-broker release from Github release.
This updates to the latest release of rds-broker. This is actually a
no-op change because there were some changes introduced to the
rds-broker[1] that were then reverted[2]. It's worth updating so that
the pipeline is up to date, making future changes easier.

This also changes the way we consume the release. We were previously
building a dev release in the pipeline from the git sources. This is
somewhat inefficient, and also has issues with the way that the version
is pinned in the git resource, especially when we have to go back a
version. Building a release once and uploading to a git release makes
this much simpler, and also means that we consume the release in the
same way as 3rd-party releases that are published to bosh.io.

[1]https://github.com/alphagov/paas-rds-broker/pull/18
[2]https://github.com/alphagov/paas-rds-broker/pull/20