Perform a fly sync after login
Our fly_sync_and_login.sh file will download the fly command from the
fly server if the local file is older than the one in the server.

But if the file has been locally updated with a different version more
recently, for example because one environment has a different concourse
version, the curl command will not update the binary.

But concourse fly command provides a `fly sync` that will update the
binary if there is version mistmatch. We can execute this fly sync
command just after login.