Use syslog-forwarder with metron_agent format
The `metron_agent` syslog message[1]and the `syslog-release` message[2] are
not equivalent. Also, the syslog_release one does not include the index number
of the VM.

Beause that, when we replaced the metron_agent syslog forwarding feature
with the syslog-release/syslog-forwarder, the log logstash filters to
parse the logs from logsearch-for-cloudfoundry stopped working[3]

We patched syslog-release to add the optional property
`syslog.log_template` which will make syslog-release/syslog-forwarder
configure rsyslog to send the logs using different templates, including
the one that metron_agent uses.

In this commit we change the release to our custom release and set the
property `syslog.log_template: "metron_agent"` so that
restore our ability to parse logs.

[1] https://github.com/cloudfoundry/loggregator/blob/develop/jobs/metron_agent/templates/syslog_forwarder.conf.erb#L53
[2] https://github.com/cloudfoundry/syslog-release/blob/master/jobs/syslog_forwarder/templates/rsyslog.conf.erb#L56
[3] https://github.com/cloudfoundry-community/logsearch-for-cloudfoundry/blob/develop/src/logsearch-config/src/logstash-filters/snippets/platform.conf#L11-L16