Link to datadog UI from the dashboard
We link directly to the triggered monitor page from the summary boxes, using
the `service:foo_env` tag to filter the results in the API and UI.

Instead of linking the triggered monitors page, we link to the manage monitors
page. Otherwise, the user may get confused by non-triggered monitors with the
state of 'No Data' which get returned from the datadog api.
We've not yet decided when 'No Data' counts as triggered, and the api doesn't
seem to give us a way to get only triggered monitors directly.
We'll probably return to this in a later story.