Upload test artifacts skips when there is none
The tar command would fail if the previous tests did not generate any
test artifacts with the api calls:

    Archiving test artifacts...
    tar: *: No such file or directory
    tar: error exit delayed from previous errors

This can happen if the acceptance tests are disabled, like in
staging&prod.

To avoid that, we add some logic to skip the upload if there are no
file in the artifacts directory.