make target to restart concourse and clear volumes
Automate the procedure to restart and clear all the volumes in
concourse. We run this often enough to automate it.

For several versions of concourse, we had multiple problems mostly
related to baggageclaim volumes[1][2][3][4]

The development team in concourse says being working on it and
refactoring the logic related to volume management.

Meanwhile the problem is not completely fixed, we found that the
procedure commented here: https://github.com/concourse/concourse/issues/565#issuecomment-237845601
helps a lot. Basically it will:
 * stop concourse services and workers.
 * kill all containers.
 * clear container network.
 * clean up the volumes and images in disk.
 * remove the images and volumes references in concourse DB.
 * restart concourse.

The concourse team insisted that this is not a supported procedure and
that the bugs must be fixed, but it did work for us several times and
we consider it is a fair workaround until concourse fixes the issues.

[1] https://github.com/concourse/concourse/issues/565
[2] https://github.com/concourse/concourse/issues/737
[3] https://github.com/concourse/concourse/issues/682
[4] https://github.com/concourse/concourse/issues/507