Switch rsyslog from RELP to TCP
There is an issue in our configuration when sending logs via RELP
protocol from syslog to logstash via an ELB. If the ingestor is
temporarily disconnected, the messages aren't logged anymore.
We may investigate the root cause but in the meantime, we revert to the
default TCP configuration because it doesn't show this behaviour.