Increase the timeout to push for http_closed_test
The test http_closed_test does deploy an static app. But we set the
DEFAULT_TIMEOUT of 30s to deploy it.

Currently we run several tests in parallel, and the load might make the
operation last longer, so the test fails due timeout.

We replace the DEFAULT_TIMEOUT with the CF_PUSH_TIMEOUT, currently 2m,
and that is used for `cf push` in general.