Avoid - and _ in generated passwords.
We've recently hit an issue with a generated password that started with
a - tripping up one of the CLI tools we use (`cf
create-service-broker`).

This therefore changes the password generation scheme we use to use hex
instead. It also increases the length from 12 to 18 chars so that the
generated passwords continue to represent the same number of random
bytes.