Add adhoc bosh ssh task to clean old rsyslog conf
In #561 we added some logic to remove the old rsyslog config created
by metron_agent. But this solution did not work completely due some
reason, and some VMs keeped the old file.

As it is quite slow and tedious the approach of creating an adhoc task
to delete the file, we will use a concourse task that runs a `bosh ssh`
on all the VMs.

This is simplier, quicker to develop and review, and easier to
test and later remove. I also prevent unnecesary bosh deployments and
restart of services after removing the task.

Currently this task will do bosh ssh on each VM, check if the file
/etc/rsyslog.d/00-syslog_forwarder.conf is present, and if it is, remove
it and restart rsyslogd.

This commit can be later reverted.