Layout dashboard vertically sized to window.
This updates the Smashing datadog dashboard to arrange the 3 boxes
vertically. It also updates the sizes to be dynamic based on the window
size. In order to get the fonts sized accordingly, I've had to use the
viewport-percentage units[1]. This is the best that's possible with pure
CSS, but they will need updating if we add/remove rows.

[1] https://drafts.csswg.org/css-values/#viewport-relative-lengths