Increase delay of clean_concourse plugin to 2s
This plugin assumes that all the elements are present and the page
rendered after 10ms, and if not, it fails because it cannot find the
right DOM elements.

The new version of concourse 2.4.0 takes longer to completely render in
the dashboard screen, specially when we have 3 frames as we do. Because
that the script does not work in our dashboard.

In this commit we workaround this issue by delaying the execution of
the script much more longer, up to 2000ms or 2s.

A more elegant solution should be waiting/polling until all elements are
present and then apply the changes. But as this is a simple script for
our dashboards it is not so relevant.