UAA failed login monitors
Add 2 monitors based on UAA failed login metrics. These metrics are
described here:
https://docs.cloudfoundry.org/loggregator/all_metrics.html#uaa

Anomaly detection is used here, not to consider seasonal changes, but
rather a more adaptative change alert.

The monitor will trigger if 50% of values are outside the "normal" range
in the past 30 minutes. This means:

* It may trigger after 15 minutes of the start of a brute force attack
* When the attack is finished, the monitor may take about 10 min to
decide the new state is the new "normal"
* It will resolve 15 min later