Generate job:count tfvar from manifest
In order to configure the datadog healthy datadog monitors, we need
the number of instances declared in the manifest for each job.

To do so we add a new script, job_instances.rb, which generates
a .tfvars file with a `job_instances` variable with a list of entries
`<job_name>:<count>`. This variable will be later used by the terraform
configuration.

The script will not output jobs with 0 instances.

We also add the logic in the pipeline to generate and pass the .tfvars
file when applying datadog terraform definitions.