Create datadog bosh job healthy monitors
Create a datadog monitor for the number of healthy bosh jobs based
on the bosh metric `bosh.healthmonitor.system.healthy`

The metric `bosh.healthmonitor.system.healthy`, which is tagged for each
deployment and job type, will report if all the monit services are
running as expected in the given job/VM.

We create one datadog monitor for each bosh job type, and we monitor
if the sum of healthy instances for each job is below the number of
desired instances in the manifest, being Warning 1 below and critical
2 below.

In order to get the number of desired instances, we use the variable
job_instances, which is a list of pairs `<job_name>:<job_count>`.
This variable is generated by script `scripts/job_instances.rb`
by parsing the manifest.

The variable is first parsed using a `null_resource`,  which computes
and exposes the name and the `critical` and `warning` thresholds.
See https://www.terraform.io/docs/provisioners/null_resource.html

Notes:
 * In the case of jobs with only one instance, the `critical`
   threshold will be -1. That means that the monitor will never be critical
   as the number of healthy instances will never be less than 0.
   We accepted this as we assume that a VM with 1 instance is not critical.

 * If bosh or NATS are redeployed, we will not get metrics at all. The
   monitors will alert with `No data` after 30 min, to cover the time to
   redeploy bosh.

It is recommended to use terraform > 0.7.8, as it fixes this bug:
https://github.com/hashicorp/terraform/pull/9466