[pipecleaner] Don't assume dummy variables are safe.
When shellchecking scripts in the pipelines, pipecleaner inserts dummy
export statements for each of the PARAMS from the pipeline to prevent
shellcheck complaining about undefined variables. Because these were set
to the static string 'DUMMY', shellcheck wouldn't complain if they were
used unquoted in the scripts because it assumed their value was safe.
Inserting a call to `date` in these prevents shellcheck from assuming
they're safe.

With this change it's been necessary to split the definition and export
onto separate lines to prevent triggering SC2155 (Declare and assign
separately to avoid masking return values)