Fix aws-broker download URL
In https://github.com/alphagov/paas-rds-broker-boshrelease/pull/16 we
renamed the broker boshrelease from aws-broker to rds-broker. As part of
this we also renamed the github repo. This should not have been a
problem because Github automatically sets up redirects when you do
this[1]. However, these redirects are a 301 Moved Permanently redirect,
and the bosh director only currently follows 302 redirects (see [2] for
details)

[1]https://help.github.com/articles/renaming-a-repository/
[2]https://github.com/cloudfoundry/bosh/pull/1523