Use terraform to setup all bosh-lite requirements
We configure using terraform all the bosh-lite requirements: VPC, subnet,
security group and SSH keys.

As we stop using a shared network, all the VPC, subnet and SG names
are parametrised by the environment name.

We move the creation of the SSH key to terraform as well.

Terraform can be optionally disabled, so the user can specify an
alternate set of AWS resources to use. This can be useful in a CI
environment.