CF: reduce update watch time to 7 mins
We want to alert on issues. But this is tricky when BOSH is updating VMs at
the same time. To workaround false-positives, increase alert thresholds to
time longer than what it takes to update VM. And also decrease this time to
7m. Currently the slowest job to deploy/update is uaa, which takes 4.5 mins.
Next one is API, about 2.5 mins. 7 mins is thus reasonable time to finish
updating/deploying the jobs. If anything is still broken afterwards, we want
to alert.