Adjust default no_data timeout to 7m
We set update_watch_time for our deployment to 7 mins, indicating that we
expect it can take up to 7 mins for BOSH to update a VM. During this time,
it is possible that data for checks will not be available, because the
processes that produce the data (datadog agent or process emiting CF
metrics) are being re-deployed. Adjust all our `no_data_timeframe`s to 7
mins to reflect this.