Replace java-buildpack with offline version
java-offline-buildpack does version pinning in contrast to
java-buildpack that is designed to fetch dependencies from internet and
keep the deps versions as recent as possible.

Note that the name of this packages must be the name of the actual
bosh package (see in the cf-release the packages with buildpack in the
name [1]). This package name might not be the same than the bosh job. If
there is a typo or the name is wrong, it will skip it silently.

[1] https://github.com/cloudfoundry/cf-release/tree/master/packages