Remove NATS and router hosts monitors
These monitors are producing false alerts during deployments and are
mostly covered by the 'bosh job is healthy' monitors.

The reason for the false alerts is that Datadog thinks the hosts are part
of the cluster for 24 hours after Bosh destroys them. You can mute hosts
in Datadog, but not for cluster checks.

We have checks set up for Bosh jobs (instance groups, for example NATS)
which use the `bosh.healthmonitor.system.healthy` metric. These checks
cover the situation where parts of a cluster go down. However, it does
mean an increased dependence on Bosh Healthmonitor metrics, which have
a no_data_timeout of 30 minutes because it takes so long to deploy Bosh.
Mitigations for this will have to be introduced (for example, alerting
when Bosh is down or a Bosh which is more highly available).