Reintroduce haproxy for logsearch
We removed haproxy a while ago because we weren't using it.

Now we add it back, purely so that we have something to do
authentication in front of Kibana.

HAProxy does http basic auth for the inbound traffic.

The single user is shared, and the password is available in cf-secrets.yml
and deliberately not available in paas-pass, so that only team members with
a prod AWS IAM user have access.