Generate test user password
If we don't generate and specify this, the tests will revert to 'meow' as
default password. This is not only insecure, but also breaks our password
policies, making the tests fail.

Generate the password using the mechanisms we have in place already to also
enable us to rotate it more easily in case we need to.