Tests: always dump whole acceptance_tests yaml
We don't need to selectively pick which properties we'll pass into tests
config. All properties configured in yaml are meant for the tests, so simply
convert them from yaml to json and pass to test. We thus don't need a
special script to do it, as we can do it in the concourse task directly.

See
https://github.com/cloudfoundry/cf-acceptance-tests/blob/d25456f3c194fdba376fa1cc7faa919b68cbafeb/vendor/github.com/cloudfoundry-incubator/cf-test-helpers/config/config.go#L11
for option parsing by tests.