Drop login null properties
Setting something to null means that the default value will be used. Which
is the same as not setting the property at all. I could not find any
significant reason why we want to include properties in our manifest, but
not set them. Properties not being used increases chance that we won't
update them as we move to newer releases that change property names. And it
removes burden of checking for these changes.