Add monitors to check the Cell garden
We check if the process is up and if more than 50% of the garden processes
are reachable to the debug port.

Note that in the case of garden-runc the process is currently called
guardian. Also, we only monitor the debug port of garden because the
service socket is a unix socket that datadog does not currently support.