Set SSL negotiation policy for all ELBs
The default behaviour if this is unspecified is for the ELBs to get the
AWS default security policy at the time that the ELB is created. This
has led to our long-lived deployments having an older security policy,
which raises warnings in Trusted Advisor etc.

This therefore updates all ELBs that terminate SSL to use the latest
built-in AWS policy (2016-08).

We previously attempted this in 2e96ba4, but had to revert. This was due
to a misunderstanding about how these policies work. An ELB has one or
more policies defined for it. Each SSL listener then has one of these
policies attached to it. The previous commit failed because we were
attempting to attach a policy to the listeners that didn't exist on the
ELB.

This commit now defines both the policy and the attachment to avoid this
problem. We can use a built-in policy by creating a policy with the
`Reference-Security-Policy` attribute set to the relevant built-in name.

Note: we've had to include the built-in policy name in the resource
name to work around an issue with this terraform resource where it
wasn't updating the policy when the param changed (presumably because
this resource is managint to items in AWS - the policy and the
attachment, and therefore doesn't deal correctly with the policy itself
changing, only the attachment).