Link to the triggered monitor view
When working on support I want to be able to quickly see the triggered monitors
which are causing an environment to be non-green.

Previously we were linking to the "manage monitors" tab which showed all the
monitors. Now we link to the "triggered" tab that only shows Warning, Critical
and No Data monitors.