STS token creation script
This script uses the existing AWS keys in the environment to issue an
STS token that is cached on the local filesystem. This can then be
sourced by the user in order to operate against the API with MFA
authenticated credentials.

It is suggested that developers create bash aliases to both create and
source the correct credentials for an environment. Such as:

    alias fetch_token="${PAAS_CF_DIR}/scripts/create_sts_token.sh" && \
        source "${HOME}/.aws_sts_tokens/${AWS_ACCOUNT}_${DEPLOY_ENV}.sh"