Only have one env-specific YAML per env
Previously we had env-specific configuration under manifests and another under
cloud-config because our manifests were split that way.

This seemed a bit excessive because we only actually had two settings so far
that we vary between envs. They are now merged into a single directory so we
only have one place to look for them.