Add datadog-route-emitter template
We want to monitor the route-emitter processes and endpoints, so
we need to add the template datadog-route-emitter to configure datadog.

After this pull request was raised a related change was merged which
bumped the version of the datadog-for-cloudfoundry release[1]. This
means we do not need to bumpt the version here and version 0.1.3 is
skipped entirely.

[1] https://github.com/alphagov/paas-cf/pull/655/commits/fb091af1400d7fa4cbf9b06c694e3b9e48864253