acceptance-tests: Improve XFF header tests
The existing tests for individual indexes of the XFF slice made it quite
hard to understand the complete behaviour. For example, the real egress IP
is the second-to-last element rather than the last element.

They were also susceptible to false-positives because we were checking that
the second element didn't belong to a 10/8 private network, rather than
asserting the actual value of the second-to-last element. So the tests would
continue to pass if an intermediate proxy inserted more elements before our
real egress IP.

Make the tests clearer and more strict by obtaining our real egress IP and
comparing the entire contents of the header. Because we're checking the
entire contents, rather than individual elements, the two existing tests
have been combined into one. This also makes it slightly faster because we
only have to push the app once.

Unfortunately we need to use a third-party service to get the egress IP. I
chose `canhazip.com` because it's reputable (been around for a while, run by
someone well known) and the response format is simple. There's some more
information about it here:

- https://major.io/icanhazip-com-faq/

I have left a FIXME comment next to the `127.0.0.1` element. This IP is
inserted by the hop between HAproxy and gorouter. It's undesirable, because
tenants should expect to get the IP from the last element, and we hope to
remove it in the future. In the meantime it is intentionally documented by
the test.