Plot summarised CC requests rate
In the grafana dashboard we plot the CC requests rate and 4xx
and 5xx rate, so we can understand the normal traffic rate and spot rise
in errors.

The metric is generated with a resolution of 10s, but we plot a total
of 12 hours. Because that the graph is really spiky and noisy.

A solution would summarise the metric in a 5 minute window, creating
a more smooth and easy to read metric. We use 5m instead of 1m because
the requests to the API have a brust pattern during deployments and smoke
tests.

We add a plot of the maximum peak of requests per second within 10s in the
5m window.

In order to highlight errors, for 4xx and 5xx we also plot the maximum
in the 5m window.