Add racoon release
We've integrated the SAP/ipsec-release into our CF and enabled its
use for two of our nodes, to ensure encrypted communication between
them.

We're introducing IPsec between these two types of nodes because they
are one of the few communications within Cloud Foundry that don't have
any transport encryption of their own and we want to secure tenant HTTP
traffic as close to the application container as possible.