Increase number of CC workers to 3 per VM
Datadog is alerting about high number of jobs in the CF api queue[1] while
the tests run. It seems that after increase the paralelism #614 in the tests
the load on the API increased.

Checking the CPU and load of the APIs, we found that currently the load on
the CC worker VMs is:

 * api: ~50% CPU and ~0.5 load during tests
 *  api workers: ~25-30% CPU and ~0.3 load.

Currently we run:

 * api: 1 main process, 2 workers
 * api_worker: 1 workers

In this PR we increase this to:

 * api: 1 main process, 3 workers
 * api_worker: 3 workers

[1] https://app.datadoghq.com/event/event?id=3649877123650295918 both in CI as in staging.