Fix bosh-vms-test for multiple bosh deployments.
Now that concourse is deployed by bosh, it's vm appears in the output of
`bosh vms`. This VM also has more than one IP, so the output appears as
follows:

+--------------------+---------+-----+-----------+-------------+
| VM                 | State   | AZ  | VM Type   | IPs         |
+--------------------+---------+-----+-----------+-------------+
| concourse/0 (UUID) | running | n/a | concourse | 10.0.0.10   |
|                    |         |     |           | 192.0.2.134 |
+--------------------+---------+-----+-----------+-------------+

This extra mostly-blank line doesn't match the grep filter, and
therefore causes this test to fail.

We fix this by specifying the deployment name in the `bosh vms` call to
only list the vms for the CF deployment, and exclude the concourse
deployment.