Only enable routes drop monitor in dev and prod
The router_routes_drop monitor get easily triggered when running the
tests in environments with little routes, like ci or staging.

We do not need to monitor this on those environments, but it is still
useful in prod. We can also keep it in dev for testing/development.

As our current version of terraform (v0.7.x) does not have conditionals,
added on v0.8.0 [1], in order conditionally disable the monitor we use the
following workaround:

 * We define a variable to enable (1) or disable (0) the resource. In
   this case defaults to 1, enabled.

 * In the resource we use the count attribute[2] set to the value of
   the previous variable.

 * In the custom variable files from each environment, we override the
   variable, in this case ci and staging, to 0.

This hack can later be changed with the conditionals from terraform.

[1] https://www.terraform.io/docs/configuration/interpolation.html#conditionals
[2] https://www.terraform.io/intro/examples/count.html