CC queue length monitors the avg in last 30m
Since we increased the number of jobs in parallel in the tests in #614,
the CC queue length monitor gets triggered often, specially in CI and
staging.

But that monitor is too sensitive, as it gets triggered if any value in
the last 30 minutes is more that 6 or 10. We change this to alert when
the average on the last 30 minutes, making the monitor resilient to
spikes. We will still detect if jobs get stuck for too long.

I also increase the warning to 7 because some of the lastest runs of
acceptance tests in CI got an average in 30m above 6.