Plot summarised app requests rate
In the grafana dashboard we plot the tenant's app requests rate and 4xx
and 5xx rate, so we can understand the normal traffic rate and spot rise
in errors.

The metric is generated with a resolution of 10s, but we plot a total
of 12 hours. Because that the graph is really spiky and noisy.

A solution would summarise the metric in a 1 minute window, creating
a more smooth and easy to read metric.

To do so, we use the count of requests instead of the rate.
We sum the values from all the gorouters, and then summarise in 1m
window using sum. Finally we scale to seconds, as the metric is reported
every 10s.