Allow apostrophes in tenant email addresses
When creating a user with the email address firstname.o'lastname@domain
this script would previously report that the email address was invalid.

The regex used to validate email addresses doesn't allow all email
addresses allowed by RFC-822, but it's apparently hard to make a concise
regex that does.[1] So in this case I added apostrophe to the regex.

I couldn't get `aws ses send-email` to accept an email address with an
apostrophe in it when using the `--destination ToAddresses=string`, so I
changed to specifying the destination email addresses as JSON.

[1] http://www.ex-parrot.com/~pdw/Mail-RFC822-Address.html