Fix regex on Linux
In e25789c I amended the regex to allow ' in email addresses. It turned out
that this broke the script on Linux because of differences in the underlying
regex engines on Mac OS X and Linux.

On Linux, to match the '-' character, it must be the last character in the
character group, otherwise it's treated as part of a range eg. as in 'A-Z'. It
seems that it failed to parse the regex because it didn't understand the range
+-\ and the bash =~ operator returned 2.

Mac OS X accepted the regex, although I'm not clear whether it created the
range according to the colleation sequence or just ignored it and used literals.

I've looked at re_format(7) on both OSes and found no obvious explanation for
the difference.